Revert "moving PreferenceScopeService behind an interface"
This reverts commit e5a21a8bd9cb4d1a09a51f110b584b9426b07c26.
