+++ lib/Wikia/src/Service/User/Preferences/Migration/PreferenceScopeService.php
@@ -2,8 +2,50 @@
 
 namespace Wikia\Service\User\Preferences\Migration;
 
-interface PreferenceScopeService {
-	public function splitLocalPreference( $option );
-	public function isLocalPreference( $option );
-	public function isGlobalPreference( $option );
+class PreferenceScopeService {
+
+	const GLOBAL_SCOPE_PREFS = 'global_scope_preferences';
+	const LOCAL_SCOPE_PREFS = 'local_scope_preferences';
+
+	/** @var array */
+	private $globalPreferenceLiterals = [];
+
+	/** @var string */
+	private $globalPreferenceRegex;
+
+	/** @var string */
+	private $localPreferenceRegex;
+
+	/**
+	 * @Inject({
+	 *    Wikia\Service\User\Preferences\Migration\PreferenceScopeService::GLOBAL_SCOPE_PREFS,
+	 *    Wikia\Service\User\Preferences\Migration\PreferenceScopeService::LOCAL_SCOPE_PREFS})
+	 * @param array $globalPreferences
+	 * @param array $localPreferences
+	 */
+	public function __construct( array $globalPreferences, array $localPreferences ) {
+		$this->globalPreferenceLiterals = $globalPreferences['literals'];
+		$this->globalPreferenceRegex = '/' . implode( '|', $globalPreferences['regexes'] ) . '/';
+		$this->localPreferenceRegex = '/' . implode( '|', $localPreferences['regexes'] ) . '/';
+	}
+
+	public function splitLocalPreference( $option ) {
+		if ( preg_match( '/(.*?)-([0-9]+)$/', $option, $matches ) && count( $matches ) == 3 ) {
+			return [$matches[1], $matches[2]];
+		}
+
+		return [null, null];
+	}
+
+	public function isLocalPreference( $option ) {
+		return preg_match( $this->localPreferenceRegex, $option ) > 0;
+	}
+
+	public function isGlobalPreference( $option ) {
+		if ( in_array( $option, $this->globalPreferenceLiterals ) ) {
+			return true;
+		}
+
+		return preg_match( $this->globalPreferenceRegex, $option ) > 0;
+	}
 }
